Move service report storing into the loop
Was trying to save $serviceReportPath before $recommendation was initialized.
This ends up storing the uploaded file multiple times if multiple items are resolved with a single service report. Since the file name hash should be the same, shouldn't be a big problem. Should still try to figure out how to avoid this.

Should fix #19