Populate Add report form with survey IDs
Give user a quick way to select a survey ID to upload a survey
report for.
Survey IDs for the current year are retrieved and used to populate
the `surveyId` input form element via an HTML 5 `datalist`. User can
still enter a survey ID that isn't in the list.

Fixes #84