Get simple survey count chart working.
Every year is the same colour and it's all on the same column chart.  Not
sure I can get Consoletvs/Charts 7 to render each year as a separate
graph.

Having all the graphs on the same year does show how dramatically the
number of surveys each month has increased though.  An interesting
side-effect of the presentation.