Multiple enhancements to make importing more functional
If there are multiple tubes associated with the machine, list them and
ask the user which tube to associate the generator data with.

Make the output look nicer.

break doesn't work for skipping iterations.  Need to use continue.

Check to see if generator data for the survey ID exists already and
exit if it does.